+++ probe-rs/src/architecture/riscv/exception_handling.rs
