+++ probe-rs/src/architecture/riscv/mod.rs
@@ -23,7 +23,7 @@ pub use registers::PC;
 pub(crate) mod assembly;
 pub mod communication_interface;
 mod dtm;
-pub(crate) mod exceptions;
+pub(crate) mod exception_handling;
 pub mod sequences;
 
 /// A interface to operate RISC-V cores.
